requirements.txt
alt-profanity-check==1.2.2
aniso8601==7.0.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
Deprecated==1.2.14
docopt==0.6.2
filelock==3.9.0
Flask==2.1.3
Flask-GraphQL==2.0.1
fonttools==4.46.0
fsspec==2023.4.0
gcloud==0.18.3
googleapis-common-protos==1.61.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jwcrypto==1.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.24.3
oauth2client==4.1.3
packaging==23.2
Pillow==9.3.0
pipreqs==0.4.13
profanity-check==1.0.3
promise==2.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.1.0
regex==2023.10.3
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
Rx==1.6.3
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
sympy==1.12
textblob==0.17.1
threadpoolctl==3.2.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.18
Werkzeug==2.2.2
wrapt==1.16.0
yarg==0.1.9
